% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bscat.R
\name{bscat}
\alias{bscat}
\title{Simulation average backscattering over orienation and length by euphausiid and copopod}
\usage{
bscat(para, misc, app = FALSE)
}
\arguments{
\item{para}{Model Parameters}

\item{app}{= FALSE [boolean] function call from shiny interface or command line}
}
\value{
list with all parameters for DWBA
}
\examples{

}
\author{
Sven Gastauer
}
