% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildpos.R
\name{buildpos}
\alias{buildpos}
\title{Build shape vector of the hypothetical taget}
\usage{
buildpos(para, disp_prof = 1)
}
\arguments{
\item{para}{a list containining the shape parameters}

\item{disp_prof}{if 1 a ggplot of the shape is added to the output list, otherwise, no plot is produces}

\item{para$shape$ni}{Number of integration points}

\item{para$shape$order}{Tapering order}

\item{para$shape$rho_L}{rho = radius of curvature}

\item{para$shape$prof_name}{Custom shape, -1 for regular shape, filename for custom shape}
}
\description{
Builds all the coordinate vectors related to the shape positions of the target
body axis is along the z-axis
construct shape coordinates
a list containing the shape coordinates: r_pos (position vector), the_tilt , dr, gamma_t, taper, x, z, and plot if disp_prof=1
}
\examples{
para$shape$L <-30
para$shape$L_a <- 16
para$shape$rho_L <- 3
para$shape$order <- 7
}
