% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_profile.R
\name{create_profile}
\alias{create_profile}
\title{create_profile}
\usage{
create_profile(g, h, seg_no, corrL, filename)
}
\arguments{
\item{g}{Density contrast g}

\item{h}{Soundspeed contrast h}

\item{seg_no}{Number of segments}

\item{corrL}{Length Correlation factor}

\item{filename}{path and name .RData file to be saved}
}
\value{
Returns a list with all g, h, segment number and length correlation information, which can be used as an input to the material properties of the DWBA model
}
\description{
Create a target material properties profile file \cr

Generate a profile based on a list of g and h values, number of segments and a length correlation
}
\examples{
create_profile(h = c(rep(1.0281,28),
                     rep(1.0296,29),
                     rep(1.0288,29),
                     rep(1.0292,29),
                     rep(1.0256,29),
                     rep(1.0264,29),
                     rep(1.0275,27)),
               g = c(rep(1.0359,,28),
                     rep(1.0375,29),
                     rep(1.0367,29),
                     rep(1.0371,29),
                     rep(1.0332,29),
                     rep(1.0341,29),
                     rep(1.0353,27)),
               seg_no =7,
               corrL = 20,
               filename="profile1")
}
