% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_config.R
\name{createParaDat}
\alias{createParaDat}
\title{Save the parameter file}
\usage{
createParaDat(para, fn)
}
\arguments{
\item{para}{Parameters}

\item{fn}{Filename with path, to where the dat file should be written}
}
\value{
Returns a a .dat file with all model parameters
}
\description{
Save the model parameter. Keep standard dat format, as used in th eoriginal Matlab code, to keep maximum compatibility with previously generated files
}
\examples{
fn <- 'config_0.dat'
para <- read_para(fn)
createParaDat(para,fn)
}
