% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shape.R
\name{densify}
\alias{densify}
\title{Densify points along axis}
\usage{
densify(xy, n = 20)
}
\arguments{
\item{xy}{dataframe with xy coordinates}

\item{b}{densify factor}
}
\description{
Densify points along axis
}
