% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWBAbscat.R
\name{DWBAscat2}
\alias{DWBAscat2}
\title{simulation average backscattering over orienation and length}
\usage{
DWBAscat2(para, misc, app = FALSE)
}
\arguments{
\item{para}{Model Parameters}

\item{misc}{misc parameters}
}
\value{
list with all parameters for DWBA
}
\description{
simulation average backscattering over orienation and length
}
\examples{
.          4
}
