% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video2profile.R
\name{generate_pos}
\alias{generate_pos}
\title{Read the configuration file
Keep standard dat format, as used in Matlab, to keep maximum compatibility with previous files}
\usage{
generate_pos(x = sub$x, y = sub$y, z = sub$z)
}
\arguments{
\item{fn}{Filename}
}
\value{
shape
}
\description{
Read the configuration file
Keep standard dat format, as used in Matlab, to keep maximum compatibility with previous files
}
\examples{
sub <- shapes[which(shapes$ID==unique(shapes$ID)[x]),]
generate_pos(sub$x,sub$y,sub$z)
}
