% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shape.R
\name{get_int}
\alias{get_int}
\title{Get polygon points from contour coordinates}
\usage{
get_int(xyP, lout = 1000)
}
\arguments{
\item{xyP}{contour points with columns x, y}

\item{lout}{length of the output data frame}
}
\description{
Get polygon points from contour coordinates
}
