% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shape.R
\name{get_mid}
\alias{get_mid}
\title{Generate a ZooScatR shape file from x,y contour points}
\usage{
get_mid(fn, outdir = "", ndens = 500, prec = 2, rot = FALSE, res = 0.02)
}
\arguments{
\item{fn}{character, input filename containing 2 columns with the x, y coordinates}

\item{outdir}{character, folder to which the the output shape should be saved, defaults to '', which selectds the current working directory}

\item{ndens}{integer, number of points that will be used as a densification factor, defaults 500}

\item{prec}{decimal points precision at which duplicated point will be averaged, defaults to 2}

\item{rot}{boolean, TRUE/FALSE, used if x and y columns in the input file are flipped}

\item{res}{numeric, defines the output resolution of the shape, defaults 0.02}
}
\description{
Generates a shape file tha tcan be  used directly in a ZooScatR model, based on contour coordinates
}
