% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inhom_gh.R
\name{inhom_gh}
\alias{inhom_gh}
\title{inhom_gh}
\usage{
inhom_gh(para)
}
\arguments{
\item{para}{set of parameters}
}
\value{
Returns a list containing the constructed g, h, Cb, FlucVal (fluctuation value), ghplot (a plot of the newly generated g and h), flucplot (a plot of the fluctuation function), gh.sum (a summary of the new gh values)
}
\description{
construct inhomogeneous g and h based on given parameter
}
\examples{
new_gh <- inhom_gh(para)
new_gh$flucplot
new_gh$ghplot
}
