% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_density.R
\name{K}
\alias{K}
\title{compressibility K}
\usage{
K(S, T, p)
}
\arguments{
\item{S}{Salinity in psu (~ppm)}

\item{T}{Temperature in degrees Celsius}

\item{p}{pressure in bar}
}
\description{
compressibility K
}
