% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length_ave.R
\name{length_ave}
\alias{length_ave}
\title{averaging over length}
\usage{
length_ave(ka0, ka1, f, pdf_type, paral, app = FALSE)
}
\arguments{
\item{ka0}{original ka  ka=[ <a>-3*std_a  <a>+3*std_a]}

\item{ka1}{output ka}

\item{f}{complex form function - freq.(or ka)  x  - orient. angle}

\item{pdf_type}{distribution type       1: uniform  2: Gaussian}

\item{paral}{PDF parameters: paral[1] = no. of bins for L PDF; paral[2] = 1/3 max. deviation for uniform, = std(length) for Gaussian}
}
\value{
averaged y points
}
\description{
averaging over length
}
