% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{plot_3D}
\alias{plot_3D}
\title{plot_3D function to create a 3D plot of a prolate spheroid}
\usage{
plot_3D(para, nx = 100)
}
\arguments{
\item{para}{a list of parameters as used in ZooScatR}

\item{nx}{number of point constituing circular elements}
}
\value{
a plotly plot object
}
\description{
Creates a 3D plotly plot of the prolate spheroid shape input
for ZooScatR, based on the parameters file.
}
