% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_para.R
\name{read_para}
\alias{read_para}
\title{Read the configuration file}
\usage{
read_para(fn)
}
\arguments{
\item{fn}{Filename}
}
\value{
list with all parameters for DWBA
}
\description{
Read a DWBA configuration file. Keep standard .dat format, as used in Matlab, to keep maximum compatibility with previous model definitions.

.dat files are standard text files.

Some example config files can be found in the extdata/configs folder: (\code{system.file(package="ZooScat")}), contained within the package directory.
}
\examples{
fn <- paste0(system.file(package="ZooScat"),"/extdata/configs/config_0.dat")
para <- read_para(fn)
}
