% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{set_la}
\alias{set_la}
\title{Compute L/a for given a and L}
\usage{
set_la(para, a = 2)
}
\arguments{
\item{para}{a list of parameters as used in ZooScatR}

\item{a}{maximum radius of the input shape}
}
\value{
a plotly plot object
}
\description{
Generates L/a based on a radius and L input
}
