% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{set_res}
\alias{set_res}
\title{Set number of variable to be computed by resolution}
\usage{
set_res(para, res = 1)
}
\arguments{
\item{para}{a list of parameters as used in ZooScatR}

\item{res}{the resolution of the output variable}
}
\value{
a plotly plot object
}
\description{
Creates a 3D plotly plot of the prolate spheroid shape input
for ZooScatR, based on the parameters file.
}
