% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TScal.R
\name{ts.cal}
\alias{ts.cal}
\title{Theoretical TS of calibration sphere
R script of function for band averaged TS of tungsten 
carbide spheres translated from Matlab code (Version date 31/5/05)                                
written by D. MacLennan (Marine Lab, Aberdeen)  
updated from R version written by Sascha F?ssler 03/2010
updated Sven Gastauer 04/01/2017}
\usage{
ts.cal(
  freq = seq(38, 200),
  c = 1500,
  d = 38.1,
  mat = "TC",
  water = "sw",
  rhow = NULL,
  plot = TRUE
)
}
\arguments{
\item{freq}{numeric frequencies [kHz]; default seq(38,200)}

\item{c}{numeric ambient soundspeeds [m/s]; default 1500}

\item{d}{numeric sphere diameter [cm]; default = 38.1}

\item{mat}{string material properties of the aphere Cu = Copper, TC = Tungsten-Carbide; default 'TC'}

\item{water}{string either sw= seawater or fw = fresh water; will be ignored if rhow (density of seawater is defined); default 'sw}

\item{rhow}{numeric density of ambient seawater, if NULL parameter water will be used; default NULL}

\item{plot}{boolean TRUE/FALSE if plot should be part of the output; default TRUE}
}
\value{
dataframe with columns "F","TS","ModF2","c", where F = Frequency [kHz]; TS = TS [dB]; ModF2 = ModF^2 and c soundspeed [m/s], an optional ggplot can be part of the output
}
\description{
Theoretical TS of calibration sphere
R script of function for band averaged TS of tungsten 
carbide spheres translated from Matlab code (Version date 31/5/05)                                
written by D. MacLennan (Marine Lab, Aberdeen)  
updated from R version written by Sascha F?ssler 03/2010
updated Sven Gastauer 04/01/2017
}
\examples{
ts.cal(freq=seq(90,170,by=0.1),c=1500.5,rhow=1.02509,plot=TRUE)
}
