% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_inversion.R
\name{TS.inverse.simple}
\alias{TS.inverse.simple}
\title{Simple TS inversion through minimisation of sum of squares}
\usage{
TS.inverse.simple(TS.test, TS.mod, n.nf = 3)
}
\arguments{
\item{TS.test}{Dataframe with TS values, where column names are the frequencies}

\item{TS.mod}{Dataframe with modelled TS values, where each row is a sample and column names are the frequencies, the last columns contain the parameters}

\item{n.nf}{number of factors contained within the data frame}
}
\description{
Takes data TS and a data frame containing simulated TS at known Length and Orientation to find the closest matching parameters
}
\examples{
#load data
setwd('C:\\\\Users\\\\sveng\\\\Nextcloud\\\\ts keras\\\\')
TS.mod <- readRDS("data\\\\TS_sim.RDS")
#select random row
rr<-sample(1:nrow(TS.mod),1)
TS.inv <- TS.inverse.simple(TS.test=TS.mod[rr,1:(length(TS.mod)-3)],
                           TS.mod=TS.mod,n.nf=3)
 ##########################################################
## CREATE PLOTS
##########################################################
ggplot(data=TS.inv,aes(x=L,y=theta,fill=log10(ss)))+
  geom_raster(interpolate=TRUE)+
  scale_fill_viridis(name="Log10(Sum of Squares)")+
  scale_x_continuous(expand=c(0,0))+
  xlab("Length [mm]")+
  ylab(expression(paste("Incident Angle", theta, " [",degree,"]")))+
  scale_y_continuous(expand=c(0,0))+
  theme_minimal()+
  theme(axis.text = element_text(size=16),
        axis.title = element_text(size=18),
        legend.position = "top")
}
\author{
Sven Gastauer
}
