% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSsphere.R
\name{TS.sphere}
\alias{TS.sphere}
\title{Calculate the analytical solution for a weakly scattering sphere}
\usage{
TS.sphere(f, r, a, c, h, g, rho, theta = pi)
}
\arguments{
\item{f}{Frequency in Hz (s^-1)}

\item{r}{Range in m from center of sphere}

\item{a}{Radius of sphere}

\item{c}{Soundspeed in surrounding fluid m/s}

\item{h}{Soundspeed contrast inside/surrounding fluid}

\item{g}{Density contrast inside/surrounding fluid}
}
\description{
Scattered pressure from an incident plane wave upon a fluid sphere. Based on:
Anderson, V. C. (1950). "Sound scattering from a fluid sphere."
The Journal of the Acoustical Society of America, 22(4), 426-431.
}
\examples{
f <- 200000
r <- 10
a <- 0.01
c <- 1477.4
rho <- 1026.8
g <- 1028.9/rho
h <- 1480.3/c
theta = pi #For backscatter
ts<- TS.sphere(f=200000,r,a,c,h,g, rho)
#For a range of frequencies
ts<- TS.sphere(f=seq(10,400)*1000,r,a,c,h,g, rho, theta=pi)
plot(ts~seq(10,400),xlab="Frequency [kHz]", ylab="TS [dB re m2]", type="l",ylim=c(-160,-80))
}
