% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bscat.R
\name{bscat}
\alias{bscat}
\title{Simulation average backscattering over orienation and length by euphausiid and copopod}
\usage{
bscat(para, misc, app = FALSE)
}
\arguments{
\item{para}{[list] A list containing all the model parameters (\url{../doc/parameters.html})}

\item{misc}{[list] A list containing the soundspeed of the surrounding fluid}

\item{app}{= FALSE [boolean] function call from shiny interface or command line}
}
\value{
list with all parameters for DWBA
}
\description{
Run the DWBA model for a given set of parameters
}
\examples{
#Get filename of the parameters file
fname <- paste0(system.file(package = "ZooScat"),"/extdata/configs/config_0.dat")
#Read in teh parameter
para = read_para(fname)
#Create list with soundspeed info
misc <- list()
misc$cw <- 1500
#Change some of the settings
#Set starting frequency over which to run the model
para$simu$var0 <- 38
#Set end frequency over which to run te model
para$simu$var1 <- 300
# run DWBA based on the settings defined in the parameters file
res <- bscat(para=para, misc=misc, app=FALSE)
#plot the results of the model
res$rplot

}
\author{
Sven Gastauer
}
