% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bscat.R
\name{bscat}
\alias{bscat}
\title{simulation average backscattering over orienation and length by euphausiid and copopod}
\usage{
bscat(para, misc, app = FALSE)
}
\arguments{
\item{para}{Model Parameters}
}
\value{
list with all parameters for DWBA
}
\description{
simulation average backscattering over orienation and length by euphausiid and copopod
}
\examples{

}
