% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bscat.R
\name{bscat}
\alias{bscat}
\title{Backscattering model}
\usage{
bscat(
  para,
  misc,
  app = FALSE,
  nang = NULL,
  nl = NULL,
  simOut = TRUE,
  plotOut = TRUE
)
}
\arguments{
\item{para}{[list] A list containing all the model parameters (\url{../doc/parameters.html})}

\item{misc}{[list] A list containing the soundspeed (cw) of the surrounding fluid (to compute cw visit \code{\link{c_Coppens1981}} or \code{\link{c_Leroy08}} or \code{\link{c_Mackenzie1981}})}

\item{app}{= FALSE [boolean] function call from shiny interface or command line}

\item{nang}{[integer] Number of angular ismulations, this will overwrite increment}

\item{nl}{[integer] Number of Length simulations, will overwrite increment informaiton}

\item{simOut}{= TRUE [boolean] If TRUE ysim (results for the simulated orientations) and ysimL (results for the simulated lengths based on the results of the mean simulated angle) as well as ang (simulated orientation angles) and L (simulated lengths) are added to the function output (average is kept as well), which contains all simulated data, if FALSE, only the averaged value is kept}

\item{plotOut}{= TRUE [boolean] If results plot should be produced or not}
}
\value{
list with all parameters for DWBA
}
\description{
Run the DWBA model for a given set of parameters. Visit the vignette for more details (\url{../doc/DwbaCommand.html})
}
\examples{
#Get filename of the parameters file
fname <- paste0(system.file(package = "ZooScatR"),"/extdata/configs/config_0.dat")
#Read in teh parameter
para = read_para(fname)
#Create list with soundspeed info
misc <- list()
misc$cw <- 1500
#Change some of the settings
#Set starting frequency over which to run the model
para$simu$var0 <- 38
#Set end frequency over which to run te model
para$simu$var1 <- 300
# run DWBA based on the settings defined in the parameters file
res <- bscat(para=para, misc=misc, app=FALSE)
#plot the results of the model
res$rplot

}
\author{
Sven Gastauer
}
