% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{DWBAapp}
\alias{DWBAapp}
\title{DWBA shiny app}
\usage{
DWBAapp()
}
\value{
Runs a DWBA web application
}
\description{
This is a Shiny web application for the DWBA model.
}
\section{General outputs}{


All Shape, Orientation, Material Property and Simulation settings that are used as model input can be changed manually or be defined through configuration and profile files.

Whenever the model is run, two plots will be generated and a data table will be generated.
}

\section{Output plots}{

Two plots will be displayed after each model run.
One plot is displaying the shape of the input object and a second one showing the model output versus a selected variable. Model outputs and display variable can be selected under the Simulation Tab.

\emph{Exporting plots}

Plots can be exported by right-clicking the plot and selecting save as. The size of the plots will automatically adopt to the size of the browser window.
The values of the plot can be eplored by hovering over the plot with the mouse. The value closest to the cursor location and the distance to the closest value point are being displayed below the plot.
\emph{Zooming in the model output plot}

Zooming can be achieved by drawing a box (holding left mouse key down) on the output plot. Once the box is visible, the plot will zoom into the selected area through double-clicking. Returning to the original zoom can be achieved through double clicking on the plot outside of the box area.
}

\section{Datatable output}{

The datatable can be exported as csv file by clicking the Export button above the shape plot. The number of shown datapoints per table page can be varied by chaning the Show entries option box.
}

\section{Model parameters}{

Four different types of model input parameters an be defined:
\enumerate{
      \item Shape
      \item Orientation
      \item Material Properties
      \item Simulation
   }

These different input types are organised in tabs.
Predefined configuration files which can define all settings can be loaded by clicking the Browse button under the Load Config menu at the top of the right frame.

Details about the model, the app and the different parameters are lined out in the vignettes:
\enumerate{
      \item A brief description of the DWBA model and an introduction on how to run it from the command line (\url{../doc/DWBAapp_vignette.html}
      \item A validation of the used DWBA model, comparing results to the analytical solution of a weakly scattering sphere (\url{../doc/CompareToAnalyticalSolution.html}
      \item An introduction to the app (\url{../doc/DWBAapp_vignette.html}
      \item A table containing the parameters and a brief explanation (\url{../doc/parameters.html}
      \item Some examples of model model parameters from literature#' (\url{../doc/gh2.htmll}
      \item An introduction on how to run parallel instances of the DWBA model (\url{../doc/DWBAParallel.html}
      }
}

\examples{
DWBAapp() #run the web applications
}
