% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWBAbscat.R
\name{DWBAscat2}
\alias{DWBAscat2}
\title{Computes ka}
\usage{
DWBAscat2(para, misc, app = FALSE)
}
\arguments{
\item{para}{Model Parameters}

\item{misc}{misc parameters}

\item{app}{TRUE or FALSE, defines if the function is used within a shiny app or not. If \code{app==TRUE} the progressbar inside the app will be updated.}
}
\value{
list of \code{ka}, the wavelength \code{k * } the width \code{a}; f the square-root of the orientation aveaged scattering cross-section
}
\description{
<<<<<<< HEAD
Computes ka
=======
simulation average backscattering over orienation and length
}
\examples{

>>>>>>> 96fcff5151d2be37a569b0de9bccf0f5ac743df4
}
