% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWBAbscat.R
\name{DWBAscat2}
\alias{DWBAscat2}
\title{Computes ka}
\usage{
DWBAscat2(para, misc, app = FALSE)
}
\arguments{
\item{para}{Model Parameters}

\item{misc}{misc parameters}

\item{app}{TRUE or FALSE, defines if the function is used within a shiny app or not. If \code{app==TRUE} the progressbar inside the app will be updated.}
}
\value{
list of \code{ka}, the wavelength \code{k * } the width \code{a}; f the square-root of the orientation aveaged scattering cross-section
}
\description{
Computes ka
}
\examples{
#Get filename of the parameters file
fname <- paste0(system.file(package = "ZooScat"),"/extdata/configs/config_0.dat")
#Read in teh parameter
para = read_para(fname)
#Create list with soundspeed info
misc <- list()
misc$cw <- 1500
DWBAscat2(para,misc)
}
