% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orient_ave.R
\name{orient_ave}
\alias{orient_ave}
\title{averaging over orientation}
\usage{
orient_ave(ang, f, pdf_type, paral)
}
\arguments{
\item{ang}{incident angle}

\item{f}{complex form function as a function of frequency, ka or orient. angle}

\item{pdf_type}{distribution type       1: uniform  2: Gaussian}

\item{paral}{PDF parameters: paral[1] = angle; paral[2] = range for uniform, = std for Gaussian}
}
\value{
averaged y points
}
\description{
averaging over orientation
}
\examples{
averaging over orientation
paral = orient_ave_para
f=f
pdf_type = para$orient$PDF
paral = orient_ave_para
paral=orient_ave_para
orinet_ave(ang, f, pdf_type, paral)
}
\author{
Sven Gastauer
}
