% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoother.R
\name{smoother}
\alias{smoother}
\title{Smoothing function}
\usage{
smoother(yi, n)
}
\arguments{
\item{yi}{Vector of values to be smoothed}

\item{n}{Number of points contained in the output y}
}
\value{
y vector of smoothed y points
}
\description{
Smoother function used to build the shape of the targets when axis or taper smoothing is requested.
}
\examples{
yi <- sample(1:1000,100)
n <- 4
y <- smoother(yi,n)
}
\author{
Sven Gastauer
}
